backend/requirements.txt
fastapi[standard] == 0.112.1
uvicorn[standard] == 0.30.6
sqlalchemy[standard] == 2.0.32
mysql-connector-python == 9.0.0